[ENH] SCINet v2: move nn modules to layers/_scinet
