+++ pytorch_forecasting/layers/_scinet/__init__.py
@@ -1,0 +1,15 @@
+"""
+Implementation of SCINet model blocks.
+"""
+
+from pytorch_forecasting.layers._scinet._blocks import (
+    SCIBlock,
+    SCINetCore,
+    SCITree,
+)
+
+__all__ = [
+    "SCIBlock",
+    "SCITree",
+    "SCINetCore",
+]
